Refactor power of ten conversion into one wrapper
